+++ pipeline/__init__.py
@@ -1,5 +1,5 @@
 """Initialize the app"""
 
-__version__ = "0.1.12"
+__version__ = "0.1.13"
 __title__ = "Pipeline"